move and refactor to progress_bar.py
